nit/chore: switch to using `size_t`/`tatami::NumericMatrix::index_type` as indexing type where possible, use static_cast instead of explicit cast where possible. fix: `GOT_SIGTERM` should be marked `volatile` as per C++ standard for `std::signal`. change: use `minstd_rand` instead of `default_random_engine` as PRNG engine.

+++ inst/include/fisher_scoring_steps.h
@@ -53,7 +53,7 @@ inline VectorXd fisher_scoring_qr_step(const EMB<D1> &model_matrix, const EMB<D2
 template <class D1, class D2, class D3, class D4, class D5, class D6, class D7>
 inline VectorXd fisher_scoring_qr_ridge_step(const EMB<D1> &model_matrix, const EMB<D2> &counts, const EMB<D3> &mu, const EMB<D4> &theta_times_mu,
                                              const EMB<D5> &ridge_penalty, const EMB<D6> &ridge_target, const EMB<D7> &beta_hat) {
-  const int extra = ridge_penalty.rows();
+  const size_t extra = ridge_penalty.rows();
 
   // the sqrt(n) is important to scale the ridge_penalty by the number of samples
   // i.e. pen = sum dev(y, mu) + n * b^t Lambda^t Lambda b^t
